Backend

See merge request s06-bigdata-dist-sub2/S06P22C208!74

+++ backend/src/main/java/com/ssafy/rikey/api/controller/ArticleController.java
@@ -3,17 +3,19 @@ package com.ssafy.rikey.api.controller;
 import com.ssafy.rikey.api.request.ArticleRequestDto;
 import com.ssafy.rikey.api.response.ArticleDetailResponseDto;
 import com.ssafy.rikey.api.response.ArticleResponseDto;
+import com.ssafy.rikey.api.response.UserRankingResponseDto;
+import com.ssafy.rikey.api.response.UserResponseDto;
 import com.ssafy.rikey.api.service.ArticleService;
+import com.ssafy.rikey.api.service.UserService;
+import com.ssafy.rikey.db.entity.Article;
+import com.ssafy.rikey.db.repository.ArticleRepository;
 import io.swagger.annotations.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.NoSuchElementException;
+import java.util.*;
 
 @Api(tags = "Article", value = "게시글 API")
 @CrossOrigin(origins = {"*"})
@@ -23,6 +25,8 @@ import java.util.NoSuchElementException;
 public class ArticleController {
 
     private final ArticleService articleService;
+    private final ArticleRepository articleRepository;
+    private final UserService userService;
 
     @GetMapping("/recent")
     @ApiOperation(value = "최근 게시글 조회", notes = "최근 게시글을 조회한다.")
@@ -110,6 +114,7 @@ public class ArticleController {
     @ApiResponses({
             @ApiResponse(code = 201, message = "성공"),
             @ApiResponse(code = 204, message = "게시글 작성 오류"),
+            @ApiResponse(code = 403, message = "잘못된 유저"),
             @ApiResponse(code = 500, message = "서버 오류"),
     })
     public ResponseEntity<Map<String, Object>> createArticle(
@@ -141,6 +146,7 @@ public class ArticleController {
             @ApiResponse(code = 200, message = "성공"),
             @ApiResponse(code = 204, message = "게시글 작성 오류"),
             @ApiResponse(code = 400, message = "게시글 탐색 오류"),
+            @ApiResponse(code = 403, message = "잘못된 유저"),
             @ApiResponse(code = 500, message = "서버 오류"),
     })
     public ResponseEntity<Map<String, Object>> updateArticle(
@@ -151,10 +157,15 @@ public class ArticleController {
         HttpStatus httpStatus = null;
 
         try {
-            //유저 확인 로직 필요
-            articleService.updateArticle(articleId, articleRequestDto);
-            httpStatus = HttpStatus.OK;
-            result.put("status", "SUCCESS");
+            Article article = articleRepository.getById(articleId);
+            if (article.getAuthor().getId().equals(articleRequestDto.getUserId())) {
+                articleService.updateArticle(articleId, articleRequestDto);
+                httpStatus = HttpStatus.OK;
+                result.put("status", "SUCCESS");
+            } else {
+                httpStatus = HttpStatus.FORBIDDEN;
+                result.put("status", "WRONG USER");
+            }
         } catch (IllegalArgumentException e) {
             httpStatus = HttpStatus.NO_CONTENT;
             result.put("status", "NO CONTENT");
@@ -177,16 +188,22 @@ public class ArticleController {
             @ApiResponse(code = 500, message = "서버 오류"),
     })
     public ResponseEntity<Map<String, Object>> deleteArticle(
+            @RequestBody @ApiParam(value = "유저 id") Map<String, String> body,
             @PathVariable @ApiParam(value = "게시글 id", required = true) Long articleId) {
 
         Map<String, Object> result = new HashMap<>();
         HttpStatus httpStatus = null;
 
         try {
-            //유저 확인 로직 필요
-            articleService.deleteArticle(articleId);
-            httpStatus = HttpStatus.OK;
-            result.put("status", "SUCCESS");
+            Article article = articleRepository.getById(articleId);
+            if (article.getAuthor().getId().equals(body.get("userId"))) {
+                articleService.deleteArticle(articleId);;
+                httpStatus = HttpStatus.OK;
+                result.put("status", "SUCCESS");
+            } else {
+                httpStatus = HttpStatus.FORBIDDEN;
+                result.put("status", "WRONG USER");
+            }
         } catch (RuntimeException e) {
             httpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
             result.put("status", "SERVER ERROR");
@@ -194,6 +211,40 @@ public class ArticleController {
             httpStatus = HttpStatus.BAD_REQUEST;
             result.put("status", "BAD REQUEST");
         }
+
+        return new ResponseEntity<Map<String, Object>>(result, httpStatus);
+    }
+
+    @GetMapping("/rankings/{nickname}")
+    @ApiOperation(value = "랭킹 조회", notes = "누적 칼로리, 누적 거리, 누적 시간을 기준으로 랭킹을 조회한다.")
+    @ApiResponses({
+            @ApiResponse(code = 200, message = "성공"),
+            @ApiResponse(code = 500, message = "서버 오류"),
+    })
+    public ResponseEntity<Map<String, Object>> getRankings(
+            @RequestParam @ApiParam(value = "지역", required = true) String area,
+            @PathVariable @ApiParam(value = "유저 닉네임", required = true) String nickname) {
+
+        Map<String, Object> result = new HashMap<>();
+        List<Integer> rankings = null;
+        HttpStatus httpStatus = null;
+
+        try {
+            rankings = userService.getRankings(nickname, area);
+            httpStatus = HttpStatus.OK;
+            result.put("status", "SUCCESS");
+        } catch (NoSuchElementException e) {
+            httpStatus = HttpStatus.BAD_REQUEST;
+            result.put("status", "WRONG USER");
+        } catch (RuntimeException e) {
+            e.printStackTrace();
+            httpStatus = HttpStatus.INTERNAL_SERVER_ERROR;
+            result.put("status", "SERVER ERROR");
+        }
+
+        result.put("rankingByCalorie", rankings.get(0));
+        result.put("rankingsByDistance", rankings.get(1));
+        result.put("rankingByTime", rankings.get(2));
 
         return new ResponseEntity<Map<String, Object>>(result, httpStatus);
     }
